refactor(scripts): split config-gates-schema.js per-gate
Splits the 52.88-MI config-gates-schema.js (worst MI in repo by 12+
points and the only file under 65) into per-gate sub-schema files
under .agents/scripts/lib/config/gates/, one literal per gate:

  - lint.schema.js              ( 9 LOC, MI 131.78)
  - mutation.schema.js          (12 LOC, MI 127.34)
  - maintainability.schema.js   (16 LOC, MI 117.77)
  - coverage.schema.js          (18 LOC, MI 115.02)
  - lighthouse.schema.js        (23 LOC, MI  98.07)
  - bundle-size.schema.js       (23 LOC, MI  95.46)
  - crap.schema.js              (29 LOC, MI  96.10)
  - shared.js                   (117 LOC, MI 74.15) — shared schema
                                                       fragments
  - index.js                    (34 LOC, MI 112.52) — thin aggregator
                                                       exporting
                                                       GATES_SCHEMA

The parent .agents/scripts/lib/config-gates-schema.js is deleted; its
sole consumer (config-settings-schema.js:370) is updated to import
GATES_SCHEMA from the new aggregator path. Each schema file carries
the same `/* node:coverage ignore file -- AJV schema declaration
(data-as-code) */` header the original had, so coverage and CRAP
gates don't try to assert on pure-literal modules.

All new modules score MI 74-131 — comfortably above the project's
MI ≥ 50 floor; the worst-MI signal in the repo is no longer dominated
by this expected outlier.

No behavior change: same shape, same exports, same GATES_SCHEMA
object. 96/96 schema-related tests pass (config-schema-mirror-drift,
config-settings-schema, shipped-configs-validate). 6818/6818
quick-suite tests pass.

Closes #2987
