--- .agents/scripts/lib/config-gates-schema.js
@@ -1,236 +1,0 @@
-/* node:coverage ignore file -- AJV schema declaration (data-as-code); per-gate sub-schemas are flat literals with no business logic to test */
-
-import { SHELL_INJECTION_PATTERN_STRING } from './config-schema-shared.js';
-
-/**
- * Per-gate sub-schemas for `delivery.quality.gates.<tier>` (Story #1737).
- *
- * Every gate shares the same four-field base:
- *
- *   - `enabled`      — when `false`, the checker exits 0 with a skip line.
- *   - `baselinePath` — repo-root-relative path to the gate's baseline file.
- *   - `tolerance`    — `{ kind: 'absolute' | 'percent', value: number }`.
- *   - `floors`       — workspace-keyed `{ "*": { ... } }` absolute floor object.
- *
- * Gate-specific extras (targetDirs for crap/MI, routes for lighthouse,
- * bundles for bundle-size, coveragePath for coverage) layer on top via the
- * per-gate schemas below. Split out of `config-settings-schema.js` to
- * keep the parent module under the maintainability ceiling — schema
- * literals score low on MI because they're long and flat.
- */
-
-const SAFE_STRING = {
-  type: 'string',
-  not: { pattern: SHELL_INJECTION_PATTERN_STRING },
-};
-
-const NULLABLE_NONEMPTY_SAFE_STRING = {
-  type: ['string', 'null'],
-  minLength: 1,
-  not: { type: 'string', pattern: SHELL_INJECTION_PATTERN_STRING },
-};
-
-const LIST_OR_EXTENDER_OF_STRINGS = {
-  oneOf: [
-    { type: 'array', items: { type: 'string' } },
-    {
-      type: 'object',
-      properties: {
-        append: { type: 'array', items: { type: 'string' } },
-        prepend: { type: 'array', items: { type: 'string' } },
-      },
-      additionalProperties: false,
-    },
-  ],
-};
-
-/** Object-shaped tolerance: `{ kind: 'absolute' | 'percent', value: number }`. */
-const TOLERANCE_SCHEMA = {
-  type: 'object',
-  required: ['kind', 'value'],
-  properties: {
-    kind: { type: 'string', enum: ['absolute', 'percent'] },
-    value: { type: 'number', minimum: 0 },
-  },
-  additionalProperties: false,
-};
-
-/**
- * Workspace-keyed floors object — `"*"` catch-all optional.
- *
- * Each value is a per-component floor object whose keys are the metric
- * names the gate consumes. The metric name keyset is intentionally open
- * (`additionalProperties: { type: 'number' }`) so per-kind rollup keys
- * (e.g. `p95`, `perMethod`, `min`, `p50`, `score`, `errorCount`,
- * `warningCount`) flow through without each per-gate sub-schema having
- * to enumerate them. Story #1892 / Task #1894 affirmed this contract:
- * the open-keyset shape is what unblocks the per-rollup floors that
- * land in S6.
- *
- * Story #2032 / Task #2041: `*` is no longer required. When omitted, the
- * framework-default floor (lines:90 branches:85 functions:90 for
- * coverage, MI ≥ 70, CRAP ≤ 20) is injected by the resolver
- * (`lib/config/quality.js`, Story #2125). Operators may pin a
- * project-wide `*` floor explicitly when they want a value other than
- * the framework default, or declare named non-`*` workspaces for
- * monorepo consumers.
- *
- * The legacy `paths` escape-valve (Story #2029) and its per-row
- * enforcement machinery were removed in Story #2125 after Story #2119
- * verified the per-row path was decorative — the unified gate
- * (`check-baselines.js`, Epic #1943) only enforces project-wide
- * rollup floors.
- */
-const FLOORS_SCHEMA = {
-  type: 'object',
-  additionalProperties: {
-    type: 'object',
-    additionalProperties: { type: 'number' },
-  },
-};
-
-/**
- * Per-gate `components` map — name → glob list. Defaulted to
- * `{ '*': ['**'] }` at the resolver layer (see
- * `.agents/scripts/lib/baselines/components.js`); the schema only
- * constrains the shape when an operator declares it explicitly.
- *
- * Story #1892 / Task #1894: introduced as the shared seam between the
- * reader and writer so per-component rollups + floors can land
- * independently of any one gate.
- */
-const COMPONENTS_SCHEMA = {
-  type: 'object',
-  additionalProperties: {
-    type: 'array',
-    items: { type: 'string', minLength: 1 },
-  },
-};
-
-const GATE_BASE = {
-  enabled: { type: 'boolean' },
-  baselinePath: { ...SAFE_STRING, minLength: 1 },
-  tolerance: TOLERANCE_SCHEMA,
-  floors: FLOORS_SCHEMA,
-  components: COMPONENTS_SCHEMA,
-};
-
-const LINT_GATE = {
-  type: 'object',
-  properties: { ...GATE_BASE },
-  additionalProperties: false,
-};
-
-const COVERAGE_GATE = {
-  type: 'object',
-  properties: {
-    ...GATE_BASE,
-    coveragePath: { ...SAFE_STRING, minLength: 1 },
-    // Story #2136 / Task #2142 — bounded timeout for `npm run test:coverage`.
-    // Wired into `runCapture` via `spawnSync({ timeout, killSignal })`. A
-    // SIGKILL fired by the timeout is translated to exit code 124 (the GNU
-    // `timeout` convention) so close-validation can branch on "hang" vs.
-    // "test failed". Default 600000 (10 min) resolved in `config/quality.js`.
-    timeoutMs: { type: 'integer', minimum: 1 },
-  },
-  additionalProperties: false,
-};
-
-const CRAP_GATE = {
-  type: 'object',
-  properties: {
-    ...GATE_BASE,
-    targetDirs: LIST_OR_EXTENDER_OF_STRINGS,
-    newMethodCeiling: { type: 'integer', minimum: 1 },
-    requireCoverage: { type: 'boolean' },
-    friction: {
-      type: 'object',
-      properties: { markerKey: { type: 'string', minLength: 1 } },
-      additionalProperties: false,
-    },
-    refreshTag: { ...SAFE_STRING, minLength: 1 },
-    // Story #2165 — bounded timeout for `npm run crap:update` spawned by the
-    // baseline-attribution refresh path. Mirrors `coverage.timeoutMs` (Story
-    // #2142): a SIGKILL fired at the budget boundary maps to exit code 124
-    // so the close orchestrator can flip the Story to `agent::blocked`.
-    refreshTimeoutMs: { type: 'integer', minimum: 1 },
-  },
-  additionalProperties: false,
-};
-
-const MAINTAINABILITY_GATE = {
-  type: 'object',
-  properties: {
-    ...GATE_BASE,
-    targetDirs: LIST_OR_EXTENDER_OF_STRINGS,
-    // Story #2165 — bounded timeout for `npm run maintainability:update`
-    // spawned by the baseline-attribution refresh path. Mirrors
-    // `coverage.timeoutMs` (Story #2142).
-    refreshTimeoutMs: { type: 'integer', minimum: 1 },
-  },
-  additionalProperties: false,
-};
-
-const MUTATION_GATE = {
-  type: 'object',
-  properties: {
-    ...GATE_BASE,
-    strykerConfigPath: NULLABLE_NONEMPTY_SAFE_STRING,
-  },
-  additionalProperties: false,
-};
-
-const LIGHTHOUSE_ROUTE = {
-  type: 'object',
-  required: ['path'],
-  properties: {
-    path: { type: 'string', minLength: 1 },
-    formFactor: { type: 'string', enum: ['mobile', 'desktop'] },
-  },
-  additionalProperties: false,
-};
-
-const LIGHTHOUSE_GATE = {
-  type: 'object',
-  properties: {
-    ...GATE_BASE,
-    baseUrl: NULLABLE_NONEMPTY_SAFE_STRING,
-    routes: { type: 'array', items: LIGHTHOUSE_ROUTE },
-  },
-  additionalProperties: false,
-};
-
-const BUNDLE_DECLARATION = {
-  type: 'object',
-  required: ['name', 'path', 'limit'],
-  properties: {
-    name: { type: 'string', minLength: 1 },
-    path: { type: 'string', minLength: 1 },
-    limit: { type: 'string', minLength: 1 },
-  },
-  additionalProperties: false,
-};
-
-const BUNDLE_SIZE_GATE = {
-  type: 'object',
-  properties: {
-    ...GATE_BASE,
-    bundles: { type: 'array', items: BUNDLE_DECLARATION },
-  },
-  additionalProperties: false,
-};
-
-/** Composite `delivery.quality.gates` schema — closed shape. */
-export const GATES_SCHEMA = {
-  type: 'object',
-  properties: {
-    lint: LINT_GATE,
-    coverage: COVERAGE_GATE,
-    crap: CRAP_GATE,
-    maintainability: MAINTAINABILITY_GATE,
-    mutation: MUTATION_GATE,
-    lighthouse: LIGHTHOUSE_GATE,
-    'bundle-size': BUNDLE_SIZE_GATE,
-  },
-  additionalProperties: false,
-};
